+++ src/metrics.rs
@@ -2450,7 +2450,7 @@ impl GooseAttack {
                 request_metric.response_time,
                 request_metric.coordinated_omission_elapsed > 0,
             );
-            if self.configuration.status_codes {
+            if !self.configuration.no_status_codes {
                 merge_request.set_status_code(request_metric.status_code);
             }
             if request_metric.success {
@@ -3043,8 +3043,8 @@ impl GooseAttack {
                 "".to_string()
             };
 
-            // Only build the status_code template if --status-codes is enabled.
-            let status_code_template: String = if self.configuration.status_codes {
+            // Only build the status_code template if --no-status-codes is not enabled.
+            let status_code_template: String = if !self.configuration.no_status_codes {
                 let mut status_code_metrics = Vec::new();
                 let mut aggregated_status_code_counts: HashMap<u16, usize> = HashMap::new();
                 for (request_key, request) in self.metrics.requests.iter().sorted() {